+++ app/dainclusionproofs/noninclusion.go
@@ -1,0 +1,1 @@
+package inclusion
